requirements.txt
Django
djangorestframework
drf-yasg
mysqlclient
django-environ
Pillow
djangorestframework-simplejwt